Some small tweaks, for now revert the reference breaking.
It seems that the primary problem is modules and functions. Specifically,
module.__dict__ refers to other modules, and function.globals() == module.__dict__
We can easily filter out all references from modules and functions, but that
doesn't remove the actual module dict, and it makes it a little bit hard
to figure out which module a given dict belongs to.

What we would like is to remove references *to* modules, not from them.
That and to remove function.__globals__ since it is also not helpful.