Start working on a dict-like container for memory objects.
The main goal is that MemoryObjectCollection can be much more efficient if
we combine how the objects are stored with how they are accessed.
dicts waste a fair amount of memory storing a key,value,hash tuple,
when we have the 'address' as our key and cheap hash, and
the value can then be stored inline. Which also means we don't have
the per-object memory overhead (no gc, no refcount, type pointer, etc.)