Some tweaks for supporting _load_moc
One big one, implement itervalues() as an actual iterator.
Pre-building the whole list is considerably expensive, as we
both have to allocate and realloc the list many times, but
we also have to create all of those proxy objects ahead of
time, and deal with all of that gc overhead.